gpu: flip y/z in projection
Flipping y/z is necessary to go from jagex to gl coordinates. This was
doing it via rotx by over rotating the camera so that it was looking
down at the scene upside down.

This took me a bit to figure out and I think is unintuitive. Flip y/z in
projection() instead. The new matrix is (in row-major order)

2/w      0     0      0
0     -2/h     0      0
0        0     1    -2n
0        0     1      0

Note that, as with the previous behavior, the final z values are [0,1)
instead of [0,-1) like gl uses for NDC. Since we do not use the z values
currently that does not matter.